Implement task: Move button place
Task ID: ccd091a5-1a2e-45ff-96d8-2b7f2f7b30d0
AI Implementation completed via Auto-Devs

- Trong Task details sheet, Đưa button `Open With Cursor` trong tab `Code changes` tab qua phần `Task Actions`
